+++ modules/hanna-css/src/css/Radio.css.ts
@@ -1,0 +1,15 @@
+import { css } from 'es-in-css';
+
+import { TogglerKnob } from './styles/forms.js';
+import { DEPS } from './utils/miscUtils.js';
+
+export default css`
+  ${DEPS('FormField', 'Checkbox')}
+`;
+
+// inlined by Checkbox.css.ts
+export const Radio_css = () => css`
+  @media screen {
+    ${TogglerKnob('Radio')}
+  }
+`;