+++ modules/hanna-css/src/css/RadioGroup.css.ts
@@ -1,16 +1,15 @@
 import { css } from 'es-in-css';
 
-import { TogglerGroup, TogglerKnob } from './styles/forms.js';
+import { TogglerGroup } from './styles/forms.js';
 import { DEPS } from './utils/miscUtils.js';
 
 export default css`
-  ${DEPS('FormField', 'Checkbox')}
+  ${DEPS('FormField', 'Radio', 'Checkbox')}
 `;
 
 // inlined by Checkbox.css.ts
 export const RadioGroup_css = () => css`
   @media screen {
-    ${TogglerKnob('Radio')}
     ${TogglerGroup('RadioGroup')}
   }
 `;